add Custom queries capacity to the bots
Currently there is only available the "custom filter" query. It is meant to match and process a filter request from a given custom sentence (e.g. "I want to filter nationality equals to Spain"). The idea is to add more types of queries that the user can ask and the bot can identify.