remove botInfo.properties support
Now every bot property is stored in bot.properties (both the generator properties and the actual bot properties).
Now the bot.properties file is not generated with xatkit-metamodel-simple (it is not designed to support NLP.js properties), so the old botconfiguration.properties file now is bot.properties, and it is copied to the generated bot. There must be specified the NLP.js (or DialogFlow) properties, the language, etc.