add SqlQuery workflow
Before this process was done in the defaultFallback, but this way it was impossible to navigate through other states (for printing the answer by pages, for instance). The implemented solution performs this "defaultFallback" behaviour in a new Workflow called SQL Query that, once it obtains an answer to a user question (using a natural language to SQL/Table model deployed in a server), prints the answer by pages like in the ShowData workflow.

also refactored defaultFallback package to languageModel