refactor Filters
The objective is to not overlap/mix operator and value types (e.g. a date field with operator "equals" may be considered a text field since text fields also have the "equals" operator). Also, this commit tries to simplify the intent definitions and the keys in ContextKeys to the minimum