add multilingual Entities
The entities are stored in a json file (spanish, catalan and english languages, one json file for each language), and they are read dynamically by the bot (i.e. entities are not hardcoded in Entities.java). Operator entities are always the same, but field entities depend on the table (.csv) used by the bot, so they must be filled with values and synonyms before running the bot