refactor json entities
Now the SchemaField reads the synonyms of the fields. The fields are stored in fields.json and contain the info in different languages in the same json object.

The SchemaType takes care of generating the JSONObject containing the fields classified by types, ready to be read by the chatbot (in Entities.java). Operator entities are separated from field entities since the future idea is to store them somewhere in the xatkit core library (they are always the same).

Now all language versions of the entities are in the same file (in fields.json and fieldOperator.json)