update filter workflows
When a filter is applied (through StructuredFilter or CustomFilter), the number of entries found (when applying all the current filters) is printed.

Added the possibility of removing a filter in StructuredQuery. Again, when removing a filter the number of matched entries is printed.

This number is gotten by generating the ResultSet (with all the previous filters + the new one) and getting its number of rows. Then, the ResultSet is stored in the context session, and when printing it (in GetResult workflow) it is not necessary to generate it again, simply get it from context session.