add maxEntriesToDisplay
This integer acts as a threshold to display the resultset. If the number of entries of a result set generated after a chatbot query is less or equal than this number, the result set is displayed immediately afterwards.

Also, maxEntriesToDisplay and pageLimit are now properties inside bot.properties file (they are generated in BodiGenerator.java) and can be customized.