add CustomShowFieldDistinct
This is a new custom query. To implement a custom query that performs an operation on a resultset, it has been necessary to implement a new method Statement.executeQuery with a parameter to specify the kind of operation (there will be more operations soon). The best practice has been to split the old executeQuery code into different methods to apply filters, remove fields and run operations (if any).

A new class Operation.java contains the different operations available (to pass them as a parameter to executeQuery, from the chatbot state that calls it).

A new intent "customShowFieldDistinctIntent" has been created to perform this operation (and also the necessary training sentences stored in the intents bundle).

A new entity called "fieldEntity" is a union of numericField, textualField and dateField entities. We need an entity that contains all fields to use it in this new operation (and potentially in other future operations. That is why a new method Entities.mergeEntities has been created to merge entities)

In addition to this, "sub-workflows" of StructuredQuery and CustomQuery are no longer stored in Bot.java, but in their main workflow (because there will be more "sub-workflows" in the future, and it is better to organize them like this)