update ResultSet generation
the resultset is always generated in GetResult (therefore, the ContextKey RESULT_SET is removed) A new state in GetResult (generateResultSetWithOperationState) to create a resultSet with an additional operation (for the customQueries that perform an operation on the tabular data source).

New ContextKeys: OPERATION and OPERATION_ARGS, to store them in the context session to generate the resultset with the appropriate operation in GetResult. As a consequence, Statement.executeQuery() removed, and the new method Statement.executeQuery(Operation operation, String... args) is always used (always specify the Operation, even if there is no operation, Operation.NO_OPERATION must be used)

StructuredFilter and CustomFilter have a variable resultSetNumRows to store the number of rows of the new resultset (used to determine if the resultset is displayed immediately or only stored)