add TestCustomQuery
This test allows obtaining, for a set of utterances, the detected intents, parameters and translations to english and sql. These are the updates:

bodi-generator.properties: add new property "enable_testing = false"

BodiGenerator.java: imports the src/test/java/com/xatkit/bot package. Adds the enable_testing property to the pom.xml generation

Bot.java: the XatkitBot is created in a dedicated method createBot(). The main method calls createBot(). nlpServerClient now stored in Bot.java

CodeGenerator.java: Add conditional dependency labs-bot-testing-tools (only if property enable_testing = true). Add maven-surefire-plugin and junit-jupiter dependencies

GetResult.java: nlpServerClient now is stored in Bot.java

NLPServerClient.java: Add a new method getTranslation to get only the translations to sql and to English. Isolate the code to get the response from the server in a new method getResponse. There are 2 signatures for the method, one with a parameter Statement (to add statement info to the request; used in runQuery) and one without it (to ignore the statement info in the request; used in getTranslations)

pom.xml: add labs-bot-testing-tools dependency