update BodiGenerator
Now there are 2 new classes with a main method:

GenerateBot.java: it does the same that BodiGenerator was doing: generate a bot

GenerateMultipleBots.java: it allows generating multiple bots at once. Here, the bot name, output dir and csv file name are not gotten from the bodi-generator.properties (there are multiple ones), instead, they are obtained from the directories stored in the resources folder (1 folder per bot). These folders are named with the csv name and contain the csv and optionally the fields.json file.