add user bodi-generator user interface
Now we have a new way to execute the bodi-generator, through a web-based user interface.

also add some getters/setters in SchemaField that are used by thymeleaf to use a SchemaField in a form and fill it.

add the 'categorical' attribute to the SchemaField, allowing to add main values to that field considering this condition.

add new method signatures in BodiGenerator.java for the methods tabularDataSourceToDataSchema, createBot and createBotPropertiesFile. The UI approach uses another objects to create the tabular data source, the bot and the bot.properties

add Properties.java, storing the bodi-generator.properties and the bot.properties in 2 maps, these are filled from the UI.

add BodiGeneratorProperties.java and BotProperties.java storing constant names of the properties used in different places of the code.

add the BodiGeneratorController.java, Application.java, resources/templates/bodi-generator/main.html and resources/application.properties files that shape the UI.