add ROW_NAME
a new file rowNames.json contains names or aliases to call a row (e.g. persons, politicians, etc.). This information can be used to ask questions related to the rows.

add a new intent/workflow CustomRowCount, to tell the user the number of rows of the tabular data source. To do so, a new entity rowNameEntity is defined, containing the aliases or the names of the rows. Also, new training sentences of CustomFrequentValueInField are added to allow the user to refer to the row names (e.g. What FIELD contains more ROW_NAME?)