add CustomFieldOfValue
Given a field A and a value of a field B, this operation gets the field A of all entries with field B equal to the value. An optional operator con be applied (e.g. max value, unique values, etc.)

add dateFunctionOperatorEntity, customFieldOfValueIntent, numericFunctionOperatorIntent, dateFunctionOperatorIntent, showAllIntent and showAllDistinctIntent

add FIELD_OF_VALUE operation