add the CustomRowOfValues workflow
update the UI to allow setting the 'key' attribute of a schema field

add the 'key' attribute in SchemaField. This attribute is written in the generated fields.json, so the chatbot can get it from there. It can be filled from the UI in the /bodi-generator/fields endpoint

add the keyFields list in Entities.java

add the Operation ROW_OF_VALUES in Operation.java and Statement.java