add multilingual bot support
now the main method of the generated bot (com.xatkit.bot.App) creates one bot for each available language. The available languages are those specified in the bot.languages property in config.properties. For each language, there must be a config_lang.properties containing the language-dependant properties (e.g. config_en.properties, for english) The advantage of this change is that now with only one bot application you can deploy all the bots (which are actually the same bot but in different languages), instead of having to create and execute separated applications.

update the UI to allow adding the language-dependant properties of the bot. This is achieved by adding a button for each selected language in the properties page of the UI, where you can switch between languages to fill the language-dependant properties.