update SqlEngine.java
Before, SqlEngine was creating a drillbit (that means, each bot creates a different one).

Now, there is a new project (opendata-for-all/bodi-drillbit) that creates a single drillbit and stores all the chatbot csv's. Now the chatbots connect to that existing drillbit (all chatbots to the same drillbit)

This has the advantage that the csv can be updated (in content, not in definition and/or structure) without restarting the chatbot. Another advantage is that we don't overload the system with several drillbits. Only a single one that gets the queries of a chatbot swarm.