add session loader and downloader
from the UI, now you can load or download the session, which is composed by a BodiGeneratorObjects file.
It contains the TabularDataSource, the DataSchema and the bot Properties (and other less relevant objects).

You can download the session (containing some bot customization), and load it later, so you don't lose all the previous work