add FieldOperatorValue custom query
it replaces the old CustomFilter. It does the same job, but the condition is not stored as a 'structured query' filter. Instead, the result is shown instantly.

renamed the 'equals' and the 'different' datetime operators to 'date_equals' and 'date_different', so they can be differentiated from the equivalent textual operators